Move `SelectOracle` and `PrepareOracle` to `qualtran/interfaces/`
