+++ qualtran/bloqs/chemistry/hubbard_model/qubitization/select_hubbard.py
@@ -23,7 +23,7 @@ from numpy.typing import NDArray
 from qualtran import bloq_example, BloqDocSpec, BoundedQUInt, QAny, QBit, Register, Signature
 from qualtran._infra.gate_with_registers import SpecializedSingleQubitControlledGate, total_bits
 from qualtran.bloqs.basic_gates import CSwap
-from qualtran.bloqs.block_encoding.lcu_select_and_prepare import SelectOracle
+from qualtran.bloqs.interfaces.lcu_select_and_prepare import SelectOracle
 from qualtran.bloqs.multiplexers.apply_gate_to_lth_target import ApplyGateToLthQubit
 from qualtran.bloqs.multiplexers.selected_majorana_fermion import SelectedMajoranaFermion
