+++ qualtran/bloqs/chemistry/sparse/select_bloq.py
@@ -31,7 +31,7 @@ from qualtran import (
 )
 from qualtran._infra.gate_with_registers import SpecializedSingleQubitControlledGate
 from qualtran.bloqs.basic_gates import SGate
-from qualtran.bloqs.block_encoding.lcu_select_and_prepare import SelectOracle
+from qualtran.bloqs.interfaces.lcu_select_and_prepare import SelectOracle
 from qualtran.bloqs.multiplexers.selected_majorana_fermion import SelectedMajoranaFermion
 
 if TYPE_CHECKING:
