+++ qualtran/bloqs/chemistry/thc/select_bloq.py
@@ -33,8 +33,8 @@ from qualtran import (
 )
 from qualtran._infra.gate_with_registers import SpecializedSingleQubitControlledGate
 from qualtran.bloqs.basic_gates import CSwap, Toffoli, XGate
-from qualtran.bloqs.block_encoding.lcu_select_and_prepare import SelectOracle
 from qualtran.bloqs.chemistry.black_boxes import ApplyControlledZs
+from qualtran.bloqs.interfaces.lcu_select_and_prepare import SelectOracle
 
 if TYPE_CHECKING:
     from qualtran.resource_counting import BloqCountT, SympySymbolAllocator
