+++ qualtran/bloqs/for_testing/qubitization_walk_test.py
@@ -20,7 +20,7 @@ import scipy
 from numpy.typing import NDArray
 
 from qualtran import Signature
-from qualtran.bloqs.block_encoding.lcu_select_and_prepare import PrepareOracle
+from qualtran.bloqs.interfaces.lcu_select_and_prepare import PrepareOracle
 from qualtran.bloqs.multiplexers.select_pauli_lcu import SelectPauliLCU
 from qualtran.bloqs.qubitization.qubitization_walk_operator import QubitizationWalkOperator
 from qualtran.bloqs.state_preparation import PrepareUniformSuperposition
