+++ qualtran/bloqs/mean_estimation/mean_estimation_operator.py
@@ -21,7 +21,7 @@ from numpy.typing import NDArray
 
 from qualtran import CtrlSpec, Register, Signature
 from qualtran._infra.gate_with_registers import SpecializedSingleQubitControlledGate, total_bits
-from qualtran.bloqs.block_encoding.lcu_select_and_prepare import PrepareOracle, SelectOracle
+from qualtran.bloqs.interfaces.lcu_select_and_prepare import PrepareOracle, SelectOracle
 from qualtran.bloqs.mean_estimation.complex_phase_oracle import ComplexPhaseOracle
 from qualtran.bloqs.reflections.reflection_using_prepare import ReflectionUsingPrepare
